chore: rename to `FileAsset` where it's necessary
