+++ Backend/src/Modules/Files/Files.Domain/Events/FileAssetCreatedEvent.cs
@@ -1,0 +1,5 @@
+﻿using Files.Domain.Models;
+
+namespace Files.Domain.Events;
+
+public record FileAssetCreatedEvent(FileAsset FileAsset) : IDomainEvent;
