+++ Backend/src/Modules/Files/Files.Domain/Events/FileAssetUpdatedEvent.cs
@@ -1,0 +1,5 @@
+﻿using Files.Domain.Models;
+
+namespace Files.Domain.Events;
+
+public record FileAssetUpdatedEvent(FileAsset FileAsset) : IDomainEvent;
