--- Backend/src/Modules/Files/Files.Domain/Events/FileCreatedEvent.cs
@@ -1,5 +1,0 @@
-﻿using Files.Domain.Models;
-
-namespace Files.Domain.Events;
-
-public record FileCreatedEvent(FileAsset FileAsset) : IDomainEvent;
