--- Backend/src/Modules/Files/Files.Domain/Events/FileUpdatedEvent.cs
@@ -1,5 +1,0 @@
-﻿using Files.Domain.Models;
-
-namespace Files.Domain.Events;
-
-public record FileUpdatedEvent(FileAsset FileAsset) : IDomainEvent;
